load-generator-input: Core functionality
This commit adds the load generator input and its core functionality.
It can read a configuration file and generate messages that are
printed to stdout.

This commit also begins to add a very minimal structure to the
project.

This is a first draft version, it works but can use some improvement
later on.